+++ CHANGES.rst
@@ -22,6 +22,7 @@ Development
 - [qa] Add software test infrastructure for data acquisition subsystems
 - [qa] Add tests covering the acquisition of USPTO PDF documents and drawings
 - [mw] Refactor the bootloader subsystem
+- [qa] Add tests for EPO OPS data source
 
 
 2019-11-01 0.169.3